Revert "Score Preview Images, use multi-threading to speed up response"
This reverts commit 18edfd73f546317299ea4df60653a05afa42eb1c.

There's no safe way to do this without using Celery or RQ. I've reverted
the code to keep it here for posterity.

+++ decruck/main/wagtail_hooks.py
@@ -1,15 +1,10 @@
 from io import BytesIO
-from decruck.main.models import (
-    CompositionPage, ScorePage, PreviewScoreImage,
-    PreviewScoreImageProcessingLock
-)
-from django import db
+from decruck.main.models import CompositionPage, ScorePage, PreviewScoreImage
 from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.utils.safestring import mark_safe
 from pdf2image import convert_from_bytes
-from threading import Thread
 from wagtail.contrib.modeladmin.options import (
     ModelAdmin, modeladmin_register
 )
@@ -42,41 +37,41 @@ class CompositionPageAdmin(ModelAdmin):
 modeladmin_register(CompositionPageAdmin)
 
 
-def process_score_preview_images(instance):
-    # Delete existing images if they exist
-    PreviewScoreImage.objects.filter(score=instance.pk).delete()
-
-    images = convert_from_bytes(instance.preview_score.read())
-    for idx, img in enumerate(images):
-        file_name = '{}-{:03d}.jpg'.format(
-            instance.preview_score.name, idx)
-
-        buffer = BytesIO()
-        img.save(buffer, 'JPEG')
-
-        img_file = InMemoryUploadedFile(
-            file=buffer,
-            field_name=None,
-            name=file_name,
-            size=img.tell,
-            content_type='image/jpeg',
-            charset=None,
-            content_type_extra=None
-        )
-        PreviewScoreImage.objects.create(
-            score=instance,
-            preview_score_image=img_file,
-            page_number=idx
-        )
-
-    PreviewScoreImageProcessingLock.objects.get(score=instance).delete()
-
-
 @receiver(post_save, sender=ScorePage)
 def generate_score_preview(sender, instance, created, update_fields, **kwargs):
+    print('post save called')
     if instance.preview_score_updated:
-        PreviewScoreImageProcessingLock.objects.create(score=instance)
-        # Close all DB threads before forking
-        db.connections.close_all()
-        thread = Thread(target=process_score_preview_images, args=[instance])
-        thread.start()
+        print('images are being updated')
+        if update_fields:
+            PreviewScoreImage.objects.filter(score=instance.pk).delete()
+
+        images = convert_from_bytes(instance.preview_score.read())
+
+        for idx, img in enumerate(images):
+            file_name = '{}-{:03d}.jpg'.format(
+                instance.preview_score.name, idx)
+
+            # Write the extracted image to a buffer
+            buffer = BytesIO()
+            img.save(buffer, 'JPEG')
+
+            # Instantiate an uploaded file object, and create a new object
+            img_file = InMemoryUploadedFile(
+                file=buffer,
+                field_name=None,
+                name=file_name,
+                size=img.tell,
+                content_type='image/jpeg',
+                charset=None,
+                content_type_extra=None
+            )
+            PreviewScoreImage.objects.create(
+                score=instance,
+                preview_score_image=img_file,
+                page_number=idx
+            )
+        return
+
+    return
+
+
